Simplify running refresher_spec with local secrets
The root tenant ID is a secret which is being filtered by VCR's
`define_cassette_placeholder`.

This means that when the value will be different depending on if you
have a `config/secrets.yml` file locally or not.

In order to pass specs without local secrets we have to compare with the
placeholder value but that meant that if you did have a secrets.yml the
specs would fail because the real value was present instead.

This change allows for the specs to pass both with and without a
`config/secrets.yml` file (basically saving me time from having to `mv
config/secrets.yml config/secrets.yml.bak` when I'm not actively
recording VCRs)