Remove old sytle gettext initializer and the empty pot file
As of this pull request: https://github.com/ManageIQ/manageiq/pull/20490
The plugin generator no longer generates a config/initializer for populating the
gettext domain.  Instead, we pull all of the messages from the plugins into the
core manageiq.pot/po.

This commit removes this initializer and the empty .pot file.

Note, the locale/.keep is still in the plugin generator so we can clean this up
when we no longer generate plugins with it:
https://github.com/ManageIQ/manageiq/blob/7a173240b0c4fe707e6943495de27bb3589e9bcc/lib/generators/manageiq/plugin/plugin_generator.rb#L53